device: add preliminary support for wired devices
